Now 32-bit platforms run with 64-bit data!
I had to kludge a few things, but it works great.

+++ main.c
@@ -19,7 +19,7 @@ static struct counted_string *pcmd_line;
  */
 static char *str_copy(char *dest, char *src)
 {
-    int length = strlen(src);
+    size_t length = strlen(src);
     return memcpy(dest, src, length) + length;
 }
